fix evaluateExpression on maven wrapper
The evaluateExpression method is currently broken on mvnw.
Because the maven prints always its current working directory
and if maven or the wrapper must be download the output of this is printed as well.
This fix will ensure that only the last line of the maven evaluate goal is returned.